+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add support for `OAUTH_ACCESS_REVOKED` in `WebhookNotification`s
 * Add support for dispute search by `customerId`, `disbursementDate`, and `effectiveDate`
 * Add `payerId` accessor in `PayPalAccount`
+* Add support for VCR compelling evidence dispute representment
 
 ## 2.77.0
 * Fix possible NullPointerException in UsBankAccount